cleaned up component exports
