+++ components/BarDiagram/index.js
@@ -1,5 +1,5 @@
 // Styles Imports
-import { styled } from "styled-components";
+import styled from "styled-components";
 
 export default function BarDiagram({ inputArray }) {
   if (inputArray) {
