+++ components/CategorySelector/index.js
@@ -1,5 +1,8 @@
-// Styles Imports
-import styled from "styled-components";
+// Style Imports
+import {
+  StyledCategoryForm,
+  StyledCategorySelector,
+} from "./styled.CategorySelector";
 
 // Functions and Dependencies Imports
 import { useRouter } from "next/router";
@@ -55,21 +58,5 @@ function CategorySelector({ categoryData, wordList, handleFilterWordList }) {
     </StyledCategoryForm>
   );
 }
-
-// Styles
-
-const StyledCategoryForm = styled.form`
-  background-color: var(--dark-main);
-  padding: 0 25px;
-  border-radius: var(--border-radius);
-`;
-
-export const StyledCategorySelector = styled.select`
-  background-color: inherit;
-  color: var(--dark-mode-text-color);
-  padding: 8px 10px 8px 10px;
-  border: none;
-  font-size: 1.2rem;
-`;
 
 export default CategorySelector;
