+++ components/CategorySelector/styled.CategorySelector.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+// Styles
+
+export const StyledCategoryForm = styled.form`
+  background-color: var(--dark-main);
+  padding: 0 25px;
+  border-radius: var(--border-radius);
+`;
+
+export const StyledCategorySelector = styled.select`
+  background-color: inherit;
+  color: var(--dark-mode-text-color);
+  padding: 8px 10px 8px 10px;
+  border: none;
+  font-size: 1.2rem;
+`;
