+++ components/EditingForm/index.js
@@ -1,16 +1,17 @@
-// Styles Imports
-import styled from "styled-components";
-import { StyledCard } from "../StyledComponents/StyledCard";
+// Style Imports
+import {
+  StyledCenteredCard,
+  StyledFormButtonRight,
+  StyledFormTitle,
+  StyledSubmitButtonRight,
+  StyledWarningText,
+} from "./styled.EditingForm";
 import {
   StyledDropDown,
   StyledForm,
   StyledFormInput,
   StyledFormLabel,
-} from "../AddEntryForm";
-import {
-  StyledSecondaryButtonRight,
-  StyledSubmitButton,
-} from "../StyledComponents/StyledButtons";
+} from "../AddEntryForm/styled.AddEntryForm";
 import WrongIcon from "@/assets/icons/WrongIcon";
 
 // Components Imports
@@ -212,36 +213,4 @@ function EditingForm({
   );
 }
 
-// Styles
-
-const StyledCenteredCard = styled(StyledCard)`
-  position: fixed;
-  transform: translate(50%, -50%);
-  top: 50%;
-  right: 50%;
-  height: auto;
-`;
-
-export const StyledWarningText = styled.p`
-  margin-top: 0;
-  font-size: 0.9rem;
-`;
-
-const StyledFormTitle = styled.h2`
-  font-size: 1.3rem;
-  margin: 15px 0;
-  align-self: flex-start;
-`;
-
-const StyledFormButtonRight = styled(StyledSecondaryButtonRight)`
-  position: absolute;
-  margin: 0;
-  top: 1rem;
-  right: 0;
-`;
-
-export const StyledSubmitButtonRight = styled(StyledSubmitButton)`
-  margin-right: -1rem;
-  border-radius: 25px 0 0 25px;
-`;
 export default EditingForm;
