+++ components/FlashcardBack/index.js
@@ -1,7 +1,15 @@
 // Styles Imports
-import { styled } from "styled-components";
-import { StyledDefinition, StyledJPDefinition, StyledUl } from "../Entry";
-import { StyledCard } from "../StyledComponents/StyledCard";
+import {
+  RightAnswerButton,
+  StyledAnswerSection,
+  StyledFlashcard,
+  WrongAnswerButton,
+} from "./styled.FlashcardBack";
+import {
+  StyledDefinition,
+  StyledJPDefinition,
+  StyledUl,
+} from "../Entry/styled.Entry";
 import WrongIcon from "@/assets/icons/WrongIcon";
 import CorrectIcon from "@/assets/icons/CorrectIcon";
 
@@ -45,41 +53,5 @@ function FlashcardBack({ entry, studyList, setIsFront, databaseMutate }) {
     </StyledFlashcard>
   );
 }
-
-// Styles
-
-export const StyledFlashcard = styled(StyledCard)`
-  position: fixed;
-  transform: translate(50%, -50%);
-  top: 50%;
-  right: 50%;
-  height: auto;
-`;
-
-const StyledAnswerSection = styled.div`
-  padding: 0.5 0 0.5rem 0;
-  background-color: inherit;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-`;
-const StyledAnswerButton = styled.button`
-  display: inline-block;
-  border: none;
-  border-radius: 25px;
-  width: 80px;
-  height: 50px;
-  padding: 5px 25px;
-  box-shadow: var(--default-box-shadow);
-  color: var(--dark-mode-text-color);
-`;
-
-const RightAnswerButton = styled(StyledAnswerButton)`
-  background-color: var(--highlight-green);
-`;
-const WrongAnswerButton = styled(StyledAnswerButton)`
-  background-color: var(--highlight-red);
-`;
 
 export default FlashcardBack;
