+++ components/Header/index.js
@@ -1,5 +1,4 @@
-import { device } from "@/utils/globalValues";
-import styled from "styled-components";
+import { StyledCircle, StyledHeaderPosition } from "./styled.Header";
 
 export default function Header() {
   return (
@@ -8,33 +7,3 @@ export default function Header() {
     </StyledHeaderPosition>
   );
 }
-const StyledHeaderPosition = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  position: sticky;
-  top: -5rem;
-  backdrop-filter: blur(1px);
-  height: 150px;
-  z-index: 2;
-
-  @media ${device.tablet} {
-    height: 200px;
-    top: -8rem;
-  }
-`;
-
-export const StyledCircle = styled.div`
-  width: 300px;
-  height: 300px;
-
-  border-radius: 0 0 50% 50%/0 0 100% 100%;
-  background-color: var(--highlight-red);
-  color: var(--dark-mode-text-color);
-
-  @media ${device.tablet} {
-    width: 400px;
-    height: 400px;
-  }
-`;
