+++ components/SearchResults/index.js
@@ -1,5 +1,5 @@
 // Styles Imports
-import styled from "styled-components";
+import { StyledResultDisplayNoMargin } from "./styled.SearchResults";
 import { StyledCenteredButton } from "../StyledComponents/StyledButtons";
 
 // Components Imports
@@ -57,15 +57,5 @@ function SearchResults({
     </>
   );
 }
-
-// Styles
-
-export const StyledResultDisplay = styled.p`
-  align-self: center;
-`;
-export const StyledResultDisplayNoMargin = styled.p`
-  align-self: center;
-  margin: 0 1rem -10px 1rem;
-`;
 
 export default SearchResults;
