+++ components/SearchResults/styled.SearchResults.js
@@ -1,0 +1,11 @@
+// Styles
+
+import styled from "styled-components";
+
+export const StyledResultDisplay = styled.p`
+  align-self: center;
+`;
+export const StyledResultDisplayNoMargin = styled.p`
+  align-self: center;
+  margin: 0 1rem -10px 1rem;
+`;
