+++ components/UserData/index.js
@@ -1,8 +1,13 @@
 // Styles Imports
 import ProfileIcon from "@/assets/icons/ProfileIcon";
-import styled from "styled-components";
 import { StyledSection } from "../StyledComponents/StyledSection";
 import { StyledGhostButton } from "../StyledComponents/StyledButtons";
+import {
+  SmallerInlineFont,
+  StyledParagraphNoMargins,
+  StyledProfilePicturePlaceholder,
+  StyledUserInfo,
+} from "./styled.UserData";
 
 // Functions and Dependencies Imports
 import { signOut, useSession } from "next-auth/react";
@@ -30,34 +35,3 @@ export default function UserData() {
     </StyledSection>
   );
 }
-
-// Styles
-
-const SmallerInlineFont = styled.span`
-  font-size: 1rem;
-`;
-
-const StyledProfilePicturePlaceholder = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%;
-  width: 130px;
-  height: 130px;
-  border: 4px solid var(--dark-main);
-`;
-
-const StyledUserInfo = styled.div`
-  font-size: 1.5rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 5px;
-`;
-
-export const StyledParagraphNoMargins = styled.p`
-  margin: 0;
-  text-align: ${(props) => (props.$isCentered ? "center" : "left")};
-  margin-bottom: ${(props) => (props.$needsMarginBottom ? "0.5rem" : "0")};
-`;
